Add commit template
It is neccessary to get to a new level of code dev.
Commiting changes every day so team can be up to date with commits.
No comments given to previous changes due to the fact that it is
genuinely impossible to track all the mess that happened to this day.

fix Examples/CWave.py, restricted alpha values to [0, pi/2) region